Refactoring gui to be reusable
