+++ src/bin/gui.rs
@@ -1,7 +1,7 @@
 use eframe::egui;
 
 use patzer::app_state::AppState;
-use patzer::widget::ChessBoard;
+use patzer::windows::PatzerApp;
 
 pub fn main() {
     let options = eframe::NativeOptions {
@@ -10,45 +10,8 @@ pub fn main() {
         ..Default::default()
     };
 
-    let mut state = AppState::new();
-    let app = PatzerApp::new(state.clone());
-    state.start_computer_players();
+    let state = AppState::new();
+    let app = PatzerApp::new(state);
 
     eframe::run_native("Patzer Chess", options, Box::new(|_cc| Box::new(app)));
-}
-
-struct PatzerApp {
-    state: AppState,
-}
-
-impl PatzerApp {
-    pub fn new(state: AppState) -> Self {
-        Self { state }
-    }
-}
-
-impl eframe::App for PatzerApp {
-    fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
-        ctx.request_repaint_after(std::time::Duration::from_millis(10));
-
-        egui::CentralPanel::default().show(ctx, |ui| {
-            let (white_name, black_name) = self.state.player_names();
-            let current_position = self.state.current_position();
-            let selected_square = self.state.ui_selected_square();
-            let select_fn = self.state.ui_select_fn();
-            let move_fn = self.state.ui_attempt_move_fn();
-            ui.add(egui::Label::new(black_name));
-            ui.add(ChessBoard::new(
-                current_position,
-                selected_square,
-                select_fn,
-                move_fn,
-            ));
-            ui.add(egui::Label::new(white_name));
-        });
-    }
-
-    fn on_exit(&mut self, _gl: Option<&eframe::glow::Context>) {
-        println!("Goodbye! I hope you had fun 👋");
-    }
 }
